+++ test/tools/metroconfig.test.ts
@@ -11,7 +11,11 @@ import type { MetroConfig } from 'metro';
 import * as path from 'path';
 import * as process from 'process';
 
-import { withSentryBabelTransformer, withSentryFramesCollapsed } from '../../src/js/tools/metroconfig';
+import {
+  withSentryBabelTransformer,
+  withSentryFramesCollapsed,
+  withSentryResolver,
+} from '../../src/js/tools/metroconfig';
 
 type MetroFrame = Parameters<Required<Required<MetroConfig>['symbolicator']>['customizeFrame']>[0];
 
@@ -106,6 +110,191 @@ describe('metroconfig', () => {
       expect(config.transformer?.babelTransformerPath).toBe(
         require.resolve('../../src/js/tools/sentryBabelTransformer'),
       );
+    });
+  });
+
+  describe('withSentryResolver', () => {
+    let originalResolverMock: any;
+
+    // @ts-expect-error Can't see type CustomResolutionContext
+    let contextMock: CustomResolutionContext;
+    let config: MetroConfig = {};
+
+    beforeEach(() => {
+      originalResolverMock = jest.fn();
+      contextMock = {
+        resolveRequest: jest.fn(),
+      };
+
+      config = {
+        resolver: {
+          resolveRequest: originalResolverMock,
+        },
+      };
+    });
+
+    describe.each([
+      ['new Metro', false, '0.70.0'],
+      ['old Metro', true, '0.67.0'],
+    ])(`on %s`, (description, oldMetro, metroVersion) => {
+      beforeEach(() => {
+        jest.resetModules();
+        // Mock metro/package.json
+        jest.mock('metro/package.json', () => ({
+          version: metroVersion,
+        }));
+      });
+
+      test('keep Web Replay when platform is web and includeWebReplay is true', () => {
+        const modifiedConfig = withSentryResolver(config, true);
+        resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'web');
+
+        ExpectToBeCalledWithMetroParameters(originalResolverMock, contextMock, '@sentry/replay', 'web');
+      });
+
+      test('removes Web Replay when platform is web and includeWebReplay is false', () => {
+        const modifiedConfig = withSentryResolver(config, false);
+        const result = resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'web');
+
+        expect(result).toEqual({ type: 'empty' });
+        expect(originalResolverMock).not.toHaveBeenCalled();
+      });
+
+      test('keep Web Replay when platform is android and includeWebReplay is true', () => {
+        const modifiedConfig = withSentryResolver(config, true);
+        resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'android');
+
+        ExpectToBeCalledWithMetroParameters(originalResolverMock, contextMock, '@sentry/replay', 'android');
+      });
+
+      test('removes Web Replay when platform is android and includeWebReplay is false', () => {
+        const modifiedConfig = withSentryResolver(config, false);
+        const result = resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'android');
+
+        expect(result).toEqual({ type: 'empty' });
+        expect(originalResolverMock).not.toHaveBeenCalled();
+      });
+
+      test('removes Web Replay when platform is android and includeWebReplay is undefined', () => {
+        const modifiedConfig = withSentryResolver(config, undefined);
+        const result = resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'android');
+
+        expect(result).toEqual({ type: 'empty' });
+        expect(originalResolverMock).not.toHaveBeenCalled();
+      });
+
+      test('keep Web Replay when platform is undefined and includeWebReplay is null', () => {
+        const modifiedConfig = withSentryResolver(config, undefined);
+        resolveRequest(modifiedConfig, contextMock, '@sentry/replay', null);
+
+        ExpectToBeCalledWithMetroParameters(originalResolverMock, contextMock, '@sentry/replay', null);
+      });
+
+      test('keep Web Replay when platform is ios and includeWebReplay is true', () => {
+        const modifiedConfig = withSentryResolver(config, true);
+        resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'ios');
+
+        ExpectToBeCalledWithMetroParameters(originalResolverMock, contextMock, '@sentry/replay', 'ios');
+      });
+
+      test('removes Web Replay when platform is ios and includeWebReplay is false', () => {
+        const modifiedConfig = withSentryResolver(config, false);
+        const result = resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'ios');
+
+        expect(result).toEqual({ type: 'empty' });
+        expect(originalResolverMock).not.toHaveBeenCalled();
+      });
+
+      test('removes Web Replay when platform is ios and includeWebReplay is undefined', () => {
+        const modifiedConfig = withSentryResolver(config, undefined);
+        const result = resolveRequest(modifiedConfig, contextMock, '@sentry/replay', 'ios');
+
+        expect(result).toEqual({ type: 'empty' });
+        expect(originalResolverMock).not.toHaveBeenCalled();
+      });
+
+      test('calls originalResolver when moduleName is not @sentry/replay', () => {
+        const modifiedConfig = withSentryResolver(config, true);
+        const moduleName = 'some/other/module';
+        resolveRequest(modifiedConfig, contextMock, moduleName, 'web');
+
+        ExpectToBeCalledWithMetroParameters(originalResolverMock, contextMock, moduleName, 'web');
+      });
+
+      test('calls originalResolver when moduleName is not @sentry/replay and includeWebReplay set to false', () => {
+        const modifiedConfig = withSentryResolver(config, false);
+        const moduleName = 'some/other/module';
+        resolveRequest(modifiedConfig, contextMock, moduleName, 'web');
+
+        ExpectToBeCalledWithMetroParameters(originalResolverMock, contextMock, moduleName, 'web');
+      });
+
+      test('calls default resolver on new metro resolver when originalResolver is not provided', () => {
+        if (oldMetro) {
+          return;
+        }
+
+        const modifiedConfig = withSentryResolver({ resolver: {} }, true);
+        const moduleName = 'some/other/module';
+        const platform = 'web';
+        resolveRequest(modifiedConfig, contextMock, moduleName, platform);
+
+        ExpectToBeCalledWithMetroParameters(contextMock.resolveRequest, contextMock, moduleName, platform);
+      });
+
+      test('throws error when running on old metro and includeWebReplay is set to false', () => {
+        if (!oldMetro) {
+          return;
+        }
+
+        // @ts-expect-error mock.
+        const mockExit = jest.spyOn(process, 'exit').mockImplementation(() => {});
+        const modifiedConfig = withSentryResolver({ resolver: {} }, true);
+        const moduleName = 'some/other/module';
+        resolveRequest(modifiedConfig, contextMock, moduleName, 'web');
+
+        expect(mockExit).toHaveBeenCalledWith(-1);
+      });
+
+      type CustomResolverBeforeMetro067 = (
+        // @ts-expect-error Can't see type CustomResolutionContext
+        context: CustomResolutionContext,
+        realModuleName: string,
+        platform: string | null,
+        moduleName?: string,
+        // @ts-expect-error Can't see type CustomResolutionContext
+      ) => Resolution;
+
+      function resolveRequest(
+        metroConfig: MetroConfig,
+        context: any,
+        moduleName: string,
+        platform: string | null,
+        // @ts-expect-error Can't see type Resolution.
+      ): Resolution {
+        if (oldMetro) {
+          const resolver = metroConfig.resolver?.resolveRequest as CustomResolverBeforeMetro067;
+          // On older Metro the resolveRequest is the creater resolver.
+          context.resolveRequest = resolver;
+          return resolver(context, `real${moduleName}`, platform, moduleName);
+        }
+        return (
+          metroConfig.resolver?.resolveRequest && metroConfig.resolver.resolveRequest(context, moduleName, platform)
+        );
+      }
+
+      function ExpectToBeCalledWithMetroParameters(
+        received: CustomResolverBeforeMetro067,
+        contextMock: CustomResolverBeforeMetro067,
+        moduleName: string,
+        platform: string | null,
+      ) {
+        if (oldMetro) {
+          expect(received).toBeCalledWith(contextMock, `real${moduleName}`, platform, moduleName);
+        } else {
+          expect(received).toBeCalledWith(contextMock, moduleName, platform);
+        }
+      }
     });
   });
 });